+++ example/vercel.json
@@ -1,0 +1,1 @@
+{ "version": 2, "rewrites": [{ "source": "/(.*)", "destination": "/api" }] }
